==> 02_data_transform2neo4j.ipynb <==
import numpy as np
import pandas as pd
import os
import datetime
import csv

data_folder = '../Data'
np.int64
time_aggregation = -1 # time in minutes. Use -1 to disable the time aggregation

dtype_dict = {
    'STARTTIMEMS': np.int64,
    'ENDTIMEMS': np.int64
}

df = pd.read_csv(f'{data_folder}/2023.csv').dropna().astype(dtype_dict)
df.head()

Extract and format nodes

df_start = df[['STARTSTATIONNAME', 'STARTSTATIONLATITUDE', 'STARTSTATIONLONGITUDE']].rename(columns={'STARTSTATIONNAME': 'name',
                                                                                                     'STARTSTATIONLATITUDE': 'latitude:float',
                                                                                                     'STARTSTATIONLONGITUDE': 'longitude:float'})
df_end = df[['ENDSTATIONNAME', 'ENDSTATIONLATITUDE', 'ENDSTATIONLONGITUDE']].rename(columns={'ENDSTATIONNAME': 'name',
                                                                                             'ENDSTATIONLATITUDE': 'latitude:float',
                                                                                             'ENDSTATIONLONGITUDE': 'longitude:float'})
df_start.head()

nodes = pd.concat([df_start, df_end], ignore_index=True).drop_duplicates()#.dropna()

print(nodes.shape)

nodes["stationId:ID"] = nodes.index
nodes[":LABEL"] = "Station"
nodes = nodes[["stationId:ID", "name", "latitude:float", "longitude:float", ":LABEL"]]

nodes.head()

nodes['stationId:ID'] = nodes['stationId:ID'].astype(int)
nodes.to_csv(f'{data_folder}/2023_nodes.csv', index=False, header=False)

nodes_header = ['stationId:ID','name','latitude:float','longitude:float',':LABEL']
with open(f'{data_folder}/2023_nodes_header.csv', 'w', newline='') as csvfile:
    writer = csv.writer(csvfile)
    writer.writerow(nodes_header)

Extract and format edges

# nodes = nodes.drop(columns=[':LABEL'])
# nodes.head()


edges = pd.merge(df, nodes,
                 left_on=['STARTSTATIONNAME', 'STARTSTATIONLATITUDE', 'STARTSTATIONLONGITUDE'],
                 right_on=['name', 'latitude:float', 'longitude:float'], how='left')

edges = edges[['ENDSTATIONNAME', 'ENDSTATIONLATITUDE', 'ENDSTATIONLONGITUDE', 'STARTTIMEMS', 'STARTSTATIONLATITUDE', 'STARTSTATIONLONGITUDE', 'ENDTIMEMS', 'stationId:ID']].rename(columns={'stationId:ID': ":START_ID"})

edges = pd.merge(edges, nodes,
                 left_on=['ENDSTATIONNAME', 'ENDSTATIONLATITUDE', 'ENDSTATIONLONGITUDE'], 
                 right_on=['name', 'latitude:float', 'longitude:float'], how='left')
# edges.head()
edges = edges[[':START_ID', 
               'STARTTIMEMS', 
               'STARTSTATIONLATITUDE', 
               'STARTSTATIONLONGITUDE',
               'ENDTIMEMS', 
               'ENDSTATIONLATITUDE', 
               'ENDSTATIONLONGITUDE', 
               'stationId:ID']].rename(columns={'stationId:ID': ":END_ID", 
                                                'STARTTIMEMS':'starttimems:int', 
                                                'STARTSTATIONLATITUDE':'startlatitude:float',
                                                'STARTSTATIONLONGITUDE':'startlongitude:float', 
                                                'ENDSTATIONLATITUDE':'endlatitude:float',
                                                'ENDSTATIONLONGITUDE':'endlongitude:float',
                                                'ENDTIMEMS':'endtimems:int'})


to_date = lambda ms: datetime.datetime.utcfromtimestamp(ms/1000).strftime('%Y-%m-%dT%H:%M:%S.%f')
edges = edges.dropna()
edges['start_date:DATETIME'] = edges['starttimems:int'].apply(to_date)
edges['end_date:DATETIME'] = edges['endtimems:int'].apply(to_date)
edges[":TYPE"] = "CYCLES_TO"
edges.head()

Aggregate per x time

# I don't understand?
if time_aggregation > 0:
    df.groupby('category').size().reset_index(name='count')
    edges['starttimems:int'] = edges['starttimems:int'] // (time_aggregation * 60 * 1000)
    edges['endtimems:int'] = edges['endtimems:int'] // (time_aggregation * 60 * 1000)

print(edges.dtypes)

# dtype_mapping = {':START_ID': int, 'starttimems':int, 'endtimems':int, ':END_ID':int, ':TYPE':str}
# edges = pd.DataFrame(edges, dtype=dtype_mapping)
edges[':START_ID'] = edges[':START_ID'].astype('int64')
edges[':END_ID'] = edges[':END_ID'].astype('int64')
edges['starttimems:int'] = edges['starttimems:int'].astype('int64')
edges['endtimems:int'] = edges['endtimems:int'].astype('int64')
edges.to_csv(f'{data_folder}/2023_edges.csv', index=False, header=False)

edges_header = [':START_ID','starttimems:int', 'startlatitude:float', 'startlongitude:float','endtimems:int', 'endlatitude:float', 'endlongitude:float',':END_ID','start_date:DATETIME','end_date:DATETIME',':TYPE']
with open(f'{data_folder}/2023_edges_header.csv', 'w', newline='') as csvfile:
    writer = csv.writer(csvfile)
    writer.writerow(edges_header)

edges.head()

print(df.shape, "vs.", edges.shape)
