==> 05_predict_working_GWR.ipynb <==
import fastgwr
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from mgwr.gwr import GWR
from mgwr.sel_bw import Sel_BW

The given script train a GWR model

You need the following files correctly located:
* ../Data/2023_edges_header.csv
* ../Data/2023_edges.csv
* ../Data/2023_nodes_header.csv
* ../Data/2023_nodes.csv


!!! We don't recommend to launch this script. Everything is precomputed and you can already see all the visualisations

df = pd.read_csv('../Data/2023_edges.csv', header =None)
with open('../Data/2023_edges_header.csv', 'r') as file:
    columns = file.readline().strip().split(',')
df.columns = columns

df_nodes = pd.read_csv('../Data/2023_nodes.csv', header =None)
with open('../Data/2023_nodes_header.csv', 'r') as file:
    columns = file.readline().strip().split(',')
df_nodes.columns = columns
df_nodes

df_edges_extended = pd.merge(df,df_nodes,left_on=':START_ID',right_on='stationId:ID')
df_edges_extended.rename(columns={'latitude:float':'start_latitude', 'longitude:float':'start_longitude'},inplace=True)
df_edges_extended.drop(['name',':LABEL','stationId:ID'],inplace=True,axis=1)
df_edges_extended = pd.merge(df_edges_extended,df_nodes,left_on=':END_ID',right_on='stationId:ID')
df_edges_extended.rename(columns={'latitude:float':'end_latitude', 'longitude:float':'end_longitude'},inplace=True)
df_edges_extended = df_edges_extended[['start_latitude', 'start_longitude','end_latitude','end_longitude', ':START_ID',':END_ID', 'month', 'day', 'hour',
       'temperature', 'wind_speed', 'distance','conditions']]
df_edges_extended

good_weather = [
    'Quelques nuages', 'Partiellement nuageux', 'Nuageux', 
    'Généralement nuageux. Risque de bruine verglaçante', 'Généralement nuageux', 
    'Généralement ensoleillé', 'Fumée', 'Ensoleillé', 'Dégagé', 
    'Couvert', 'Alternance de soleil et de nuages', 
    'Alternance de soleil et de nuages. Risque d\'orages'
]

# Create a new column to flag good weather
df_edges_extended['good_weather'] = df_edges_extended['conditions'].apply(lambda x: 1 if x in good_weather else 0)

df_edges_extended

departure_counts = df_edges_extended.groupby([":START_ID","start_longitude","start_latitude","temperature","wind_speed","good_weather","month","day","hour"]).size().reset_index(name='count')
arrival_counts = df_edges_extended.groupby([":END_ID","end_longitude","end_latitude","temperature","wind_speed","good_weather","month","day","hour"]).size().reset_index(name='count')

train, test = train_test_split(departure_counts.iloc[:20000],train_size=0.8)
train_departure_variables = train[['temperature','wind_speed','good_weather','month','day','hour']].values
train_departure_locations = train[['start_longitude','start_latitude']].values
train_departure_count = train['count'].values.reshape((-1,1))
test_departure_variables = test[['temperature','wind_speed','good_weather','month','day','hour']].values
test_departure_locations = test[['start_longitude','start_latitude']].values
test_departure_count = test['count'].values.reshape((-1,1))

print(train_departure_locations.shape)
print(train_departure_count.shape)
print(train_departure_variables.shape)
print(test_departure_locations.shape)
print(test_departure_locations.shape)

model = GWR(train_departure_locations,train_departure_count,train_departure_variables,bw=0.1, kernel='bisquare')
results = model.fit()

results.summary()

prediction = model.predict(test_departure_locations, test_departure_variables)

y_prediction_gwr = prediction.predy
print(y_prediction_gwr)
np.savetxt('./prediction_gwr.csv', y_prediction_gwr, delimiter=',', fmt='%d')
